pan surface view's grid

+++ surfaceviewrxed/src/main/kotlin/com/rosberry/sample/surfaceviewrxed/presentation/system/drawing/SceneParams.kt
@@ -4,6 +4,6 @@ package com.rosberry.sample.surfaceviewrxed.presentation.system.drawing
  * @author mmikhailov on 30/03/2019.
  */
 data class SceneParams(
-        val width: Int,
-        val height: Int
+        val width: Float,
+        val height: Float
 )
